refactoring directories and files names to lowercase
