--- StuDash_home/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class StudashHomeConfig(AppConfig):
-    name = 'StuDash_home'
