+++ studash_/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for StuDash project.
+Django settings for studash_ project.
 
 Generated by 'django-admin startproject' using Django 3.1.3.
 
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'StuDash_home.apps.StudashHomeConfig',
+    'studash_app.apps.StudashappConfig',
 ]
 
 MIDDLEWARE = [
@@ -50,7 +50,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'StuDash.urls'
+ROOT_URLCONF = 'studash_.urls'
 
 TEMPLATES = [
     {
@@ -68,7 +68,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'StuDash.wsgi.application'
+WSGI_APPLICATION = 'studash_.wsgi.application'
 
 
 # Database
