+++ studash_/urls.py
@@ -1,4 +1,4 @@
-"""StuDash URL Configuration
+"""studash_ URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.1/topics/http/urls/
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from StuDash_home import views
+from studash_app import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
