+++ studash_/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for StuDash project.
+WSGI config for studash_ project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'StuDash.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'studash_.settings')
 
 application = get_wsgi_application()
