+++ studash_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class StudashappConfig(AppConfig):
+    name = 'studash_app'
