+++ studash_app/tests.py
