+++ studash_app/views.py
